[20221226] | [dONE] | Change filing organization to solve ng build primeng-library problem.
