+++ src/app/components/calendarpersian/public_api.ts
@@ -1,3 +1,10 @@
 /* بسم الله الرحمن الرحیم */
 
 export * from './calendarpersian';
+export * from './arithmetics.service';
+export * from './jdate';
+export * from './jdate-calculator.service';
+export * from './jdate-validator.service';
+export * from './simple-date.interface';
+export * from './simple-date.interface';
+export * from './persian-number.service';
